+++ backend/app/models/leave.py
@@ -1,4 +1,5 @@
 import uuid
+from decimal import Decimal
 from datetime import date, datetime
 from sqlalchemy import Date, DateTime, Numeric, String, Text, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
@@ -13,7 +14,7 @@ class LeaveRequest(Base, TimestampMixin):
     leave_type: Mapped[str] = mapped_column(String(16), nullable=False)  # annual/sick/personal/maternity/other
     start_date: Mapped[date] = mapped_column(Date, nullable=False)
     end_date: Mapped[date] = mapped_column(Date, nullable=False)
-    duration_days: Mapped[float] = mapped_column(Numeric(5, 1), nullable=False)
+    duration_days: Mapped[Decimal] = mapped_column(Numeric(5, 1), nullable=False)
     reason: Mapped[str] = mapped_column(Text, nullable=False)
     status: Mapped[str] = mapped_column(String(16), default="pending")  # pending/approved/rejected/cancelled
     approved_by: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), ForeignKey("users.id"), nullable=True)
