+++ backend/app/models/outsource.py
@@ -1,4 +1,5 @@
 import uuid
+from decimal import Decimal
 from datetime import datetime
 
 from sqlalchemy import DateTime, Integer, Numeric, String, Text, ForeignKey
@@ -43,10 +44,10 @@ class OutsourceTask(Base, TimestampMixin, SoftDeleteMixin):
     task_type: Mapped[str] = mapped_column(String(32), nullable=False)  # production, installation, design, transport
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
     quantity: Mapped[int] = mapped_column(Integer, default=1)
-    unit_price: Mapped[float] = mapped_column(Numeric(14, 2), default=0)
-    total_amount: Mapped[float] = mapped_column(Numeric(14, 2), default=0)
-    paid_amount: Mapped[float] = mapped_column(Numeric(14, 2), default=0)
-    unpaid_amount: Mapped[float] = mapped_column(Numeric(14, 2), default=0)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0)
+    total_amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0)
+    paid_amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0)
+    unpaid_amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0)
     status: Mapped[str] = mapped_column(String(32), default="pending")  # pending, in_progress, completed, settled
     assigned_to: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), ForeignKey("users.id"), nullable=True)
     expected_at: Mapped[datetime | None] = mapped_column(DateTime, nullable=True)
@@ -67,7 +68,7 @@ class OutsourcePayment(Base, TimestampMixin):
     payment_no: Mapped[str] = mapped_column(String(64), unique=True, nullable=False)
     vendor_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("outsource_vendors.id"), nullable=False)
     task_id: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), ForeignKey("outsource_tasks.id"), nullable=True)
-    amount: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
     payment_method: Mapped[str | None] = mapped_column(String(32), nullable=True)
     payee_company_name: Mapped[str | None] = mapped_column(String(255), nullable=True)
     paid_at: Mapped[datetime | None] = mapped_column(DateTime, nullable=True)
